Django-test-api/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Django==2.0.3
django-js-asset==1.2.2
django-mptt==0.9.0
djangorestframework==3.10.3
get==2019.4.13
idna==2.8
post==2019.4.13
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7